Generate the most important configs
