+++ rusty_ex/src/configs/mod.rs
@@ -1,5 +1,6 @@
 pub mod centrality;
 pub mod config_generator;
+pub mod config_solver;
 pub mod prop_formula;
 
 pub type CnfLit<T> = (T, bool);
